Object Exercise
I did the objects exercise by saying that if the name of the driver isn't already a property in the array add it and initialize the property value to one. As the 'for each' continues to iterate through the array, every time it sees the drivers name add one to the existing value at that property in the object.